taskManagement/backend/requirements.txt
Django>=3.0,<4.0
djongo==1.3.6  # Django MongoDB connector
djangorestframework>=3.12
pymongo==3.12.3
